Refactor HTMLConfiguratorBridge
Refactor HTMLConfiguratorBridge.ts into smaller, more manageable files, preserving all existing functionality.
